Update tutorial example with more robust timing
Switched the Controller over to timing that allows it to respond to messages whenever it receives them.

Switched the EV federate over to a 1 minute periodic timing (instead of timedelta) to allow it to get it ready to simulate the physics of the battery using value signals. I also restructured the code just a bit so that the physics simulation that is currently being done is clearly delineated with the hopes that this will make it easier to slot in the battery physics model Allison will be working on.