Add User Guide Example 4.1
Example 4.1 provides a demonstration of three kinds of federates: value, message, and combo.

The Battery federate is a value federate and models the EVs' batteries as resistors whose value depends on the SOC of said battery. All EV batteries are modeled in a single federate.

The Charger federate models a fleet of EV chargers that apply charging voltage to the Battery federate and and receive back the charging current from said federate. It uses this charge current to estimate the SOC of the battery and sends that as a message to the Controller federate. The Charger federate has noisy current measurements making the SOC estimate different than that of the Battery federates (which is modeling the true SOC).

The Controller federate receives the SOC estimates from the Charger federates and tells the Charger federate to terminate charging when the SOC reaches a near-full value.