Restructure conditional for sending messages in eq
I learned that Python evaluates conditions left-to-right so I can make a compound conditional that first checks to make sure the event queue is not empty before checking to see if any messages need to be sent.

Also, this restructuring fixes a bug that @phltop identified: if I delete a message and empty the queue, the old "while" conditional didn't check to see if the queue was emtpy.