Update data logging for graphs to account for imperfect communication
Previously we assumed all data was communicated between the Charger and the Controller. Specifically that all data from all federates was received on each time step. Now that we're modeling imperfect communication, this is not guaranteed. Some granted times only one message may be waiting and on some granted times there may be five.

When preparing the data for graphing it was assumed the timestep when each message was received would also collect data for all five messages. Since this is no longer the case, we just have to create unique timestep vectors for each EV and record the timesteps when we receive these messages said vectors. Five x vectors with five matching y vectors.