Add "event_triggered" flag to filter federate example to band-aid over bug
Adding this flag to the filter federate itself mucked with the internal HELICS timing such that the example runs successfully. Phil stated that all "event-driven" federates (rather than timing driving) should probably set this flag. Internally when a filter is added to HELICS, this flag is set for the filter. Which is to say, from this point on, all controllers and filter federates should have this flag set by default.

Also cleaned up some typos, logging messages and added some commented-out debugging messages.