Update Advanced Default data collection for plots to get it working on Ubuntu 20.04
For reasons I don't understand, the Advanced Default example didn't run successfully on Ubuntu 20.04. The plotting error-ed out with the x and y vectors not being equal in length. I was able to update the code so that the two vectors are guaranteed to be the same length (better code in that regard) but I'm unnerved that this example worked on my Mac but not on Ubuntu.