Iteration example in initialization and time loop
The example illustrates how to construct a simple co-simulation with
iteration both in the initialization as well as at between time steps.

Note: this example only considers value-federates, i.e. communication
with subscriptions and publications.

Some refactoring was performed to help use the same code in both
Battery and Charger federates and stored in a file called iterutils.py.
Actually, the federates themselves should be defined as a class.
However, that would potentially add too much change to the Example
structure, which we want to keep consistent and for that reason this
compromise is struck here.