Closes #54: inconsistent example results
The main issue seems to have been the random number seeds.
By making sure these were the same for all files:
- Battery.py
- Charger.py
- Controller.py
the results matched.

Additionally, the two examples were changed to run for the same amount
of time (1 week). Previously, advanced_default ran for 1 week, while
fundamental_integration ran for only a day.

Finally, there appears to have been a problem with the
fundamental_integration example, where the controller federate
successfully exited, but the charger and battery hung indefinitely.
Upon comparison with the advanced_default it was observed that there
was a slight difference in the time request in Controller.py:
- advanced_default:
        starttime = h.HELICS_TIME_MAXTIME
- fundamental_integration:
        fake_max_time = int(h.HELICS_TIME_MAXTIME)
        starttime = fake_max_time
By changing fundamental_integration to also simply have
    starttime = h.HELICS_TIME_MAXTIME
the hanging problem appears to have gone away.
This is somewhat unsatisfying since in both cases the request is far
greater than any of the actually granted times.
Therefore, should the problem appear again, it would make sense to start
investigating around this issue.