Changing controller time request
Requesting HELICS_MAX_TIME in the controller federate was leading to
unpredictable behavior where sometimes the simulation would hang
other times the controller would run for an extra time step, etc.

The time request has now been changed to the maximum time for the
simulation (total_interval) which is sufficient, reproduces the same
results, and at least at this point does not seem to suffer from the
same unpredictable behavior.