Working Matlab advanced default example
Advanced default example successfully converted and running with matlab.

WARNING: there is **a lot** of incompatibility between helics 3.2.0 and
helics 3.2.1. This example is working with helics 3.2.1. It is entirely
likely that it will **not** work with anything earlier.

A few notes:
- In the advanced_default_runner the script names need to be escaped
with double backslashes so instead of run('Battery.m') we get
run(\\'Battery.m\\')

- There is at this moment still an issue with helics.HELICS_TIME_MAXTIME
in matHELICS. for this reason a function called getHelicsMaxtTime is
used. Once the helics.HELICS_TIME_MAXTIME is fixed, this function will
no longer be necessary.

- There appears to, here too, be some instability about how the
federates are exiting, though this is very difficult to reproduce
consistently.