Added counter example with iteration is turned off
To demonstrate what the intra-timestep iteration is really doing a
version of the example is created where the iteration is turned off.
The resulting plots show that the model experiences quite a bit of
oscillation.

Note: Iteration during initialization is still maintained

The Battery.py and Charger.py scripts have also been slightely updated
so that the only difference between them and the no iteration version
is setting a single flag (and a few save names for figures).