Fix save path in no-iteration example version
Removed reference to a different folder which was causing errors.

NOTE: this means that when the example using no iteration is run it will
overwrite the local figures that may be stored locally from the
iteration run.
Not a very big deal, since these can be re-run/renamed, but just a
heads up.