Adding logging to Battery Federate
Added logging to battery federate in the fundamental default example.
This prints the initial soc as well as the battery capacity of the
initialized batteries.

This is need because it is hard-coded in the FMU example to make sure
we are running the same simulation and can therefore compare the power
plot at the end.