Updated connector federate example; WORKING!
Phil fixed the bug in the HELICS library and I fixed a bug related to my assumption that the pubs and subs will be created in the federate in the same order as if they were defined in a config file. This is a BAD assumption as helics_connector doesn't care what order the matches are made in, it just tries to make them.